% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencontracs.R
\docType{data}
\name{data120}
\alias{data120}
\title{Análisis de contratos públicos directos durante el Covid-19.}
\format{
Base de datos del ranking  120
}
\usage{
data(data120)
}
\description{
Análisis de contratos públicos directos durante el Covid-19.
}
\keyword{datasets}
